Add DialogFlow's dependency modules exclusions
Add module exclusions in the build.gradle file for the DialogFlow's dependency. The reason is to fix the gradle build errors without skipping lint's error checking process, as those corresponding libraries are already implemented in Android, so they were conflicting with the ones that DialogFlow provides as well.

Issues:
#65

Contributors:
@chrytsa