Add SQLite persistence storage implementation and Car object migration
Add SQLte logic to be able to perform Read, Write, Update and Create functions to a database in order to be able to store the car names and 'addresses', Change the SmartCar implementation from static enum to a plain object in order to be able to handle more dinamic updates whit the corresponding updates. Refactor the code base to be able to accommodate the migration from static enum to normal object.

Issues:
#80

Contributors:
@krasen