Update healthRover.ino file
Removed included library that we don't use right now.
Setting sensor reading to only stop when car distiance is more than 5 because the sensor is getting wrong readings of 4 and below.
Changed the order of setCarSpeed() and turnCar() because we need movement before we can turn the car.
Added response from server when obstacle is detected.

Contributors:
@samgun-6
@chrytsa