Change behaviour of the car when stopped and a turning command follows
When in reverse mode, then the user stops the SmartCar and afterwards tries to steer it to the left or to the right, the SmartCar is reversing towards the specified direction.
However, the expected behaviour would be that when a turning command to the left or to the right is received, while the SmartCar is stopped, the car shall then shall proceed to the specified direction with a positive speed (onwards left or right).